remove Crashlystic dependency from shared codes
