+++ Rocket.Chat/Managers/Launcher/LauncherProtocol.swift
@@ -15,6 +15,7 @@ protocol LauncherProtocol {
 final class Launcher: LauncherProtocol {
     private lazy var launchers: [LauncherProtocol] = {
         return [
+            ManagerDelegatesCoordinator(),
             PersistencyCoordinator(),
             BugTrackingCoordinator(),
             UserCoordinator()
