+++ SDKExample/Podfile
@@ -1,5 +1,4 @@
-# Uncomment the next line to define a global platform for your project
-# platform :ios, '9.0'
+platform :ios, '10.0'
 
 target 'SDKExample' do
   # Comment the next line if you're not using Swift and don't want to use dynamic frameworks
